+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Version 0.3.0
 * Support `.err!` and `.ok!` public methods.
+* Rename `.ok` to `.ok_result`. [Breaking]
+* Rename `.err` to `.err_result`. [Breaking]
 
 ## Version 0.2.0
 * Support `.rescue_with` to handle exceptions in `.run` block.